feat: Add Node.js script for displaying the film title from a SWAPI GET request
This script takes a film ID as a command line argument, constructs a SWAPI URL, and uses the 'request' module to make a GET request. It then prints the title of the film based on the response body. If an error occurs during the request, the script prints the error object.